apps/inference/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
onnxruntime==1.17.3
pillow==10.4.0
numpy==1.26.4

